refactor(CLI): removed a controller and added a config command
